feat: implementa Exercise, arreglando fallos de concepto.refs #14, #15, #16
